Update CT protocol to microfost standard
