+++ src/CombinatorialTestingFeature.ts
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 import { ClientCapabilities, InitializeParams, ServerCapabilities, StaticFeature, WorkDoneProgressOptions } from "vscode-languageclient";
-import { ExperimentalServerCapabilities } from "./protocol.slsp";
+import { CombinatorialTestingClientCapabilities, CombinatorialTestingServerCapabilities } from "./protocol/combinatorialTesting.slsp";
 
 export class CombinantorialTestingFeature implements StaticFeature {
     public SupportsCT: boolean = false;
@@ -9,19 +9,20 @@ export class CombinantorialTestingFeature implements StaticFeature {
 
     fillInitializeParams?: (params: InitializeParams) => void;
     fillClientCapabilities(capabilities: ClientCapabilities): void {
-        if (!capabilities.experimental)
-            capabilities.experimental = { combinatorialTesting: true };
-        else
-            Object.assign(capabilities.experimental, { combinatorialTesting: true });
+        capabilities.experimental = capabilities.experimental || {};
+        let ctCapabilities = capabilities as CombinatorialTestingClientCapabilities;
+        ctCapabilities.experimental.combinatorialTesting = true;
     }
-    initialize(capabilities: ServerCapabilities<ExperimentalServerCapabilities>): void {
+    initialize(capabilities: ServerCapabilities): void {
+        let ctCapabilities = (capabilities as CombinatorialTestingServerCapabilities);
+
         // If server supports CT
-        if (capabilities?.experimental?.combinatorialTestProvider) {
+        if (ctCapabilities?.experimental?.combinatorialTestProvider) {
             this.SupportsCT = true;
 
             // Check if support work done progress
-            if (WorkDoneProgressOptions.hasWorkDoneProgress(capabilities?.experimental?.combinatorialTestProvider))
-                this.SupportsCTWorkDoneProgress = capabilities?.experimental?.combinatorialTestProvider.workDoneProgress
+            if (WorkDoneProgressOptions.hasWorkDoneProgress(ctCapabilities?.experimental?.combinatorialTestProvider))
+                this.SupportsCTWorkDoneProgress = ctCapabilities?.experimental?.combinatorialTestProvider.workDoneProgress
         }
     }
     dispose(): void {
